chore: reorganize more test stuff
