+++ tests/e2e/cogs/customization_test.py
@@ -1,4 +1,3 @@
-import discord  # noqa: F401 Is used in file
 import pytest
 
 pytestmark = pytest.mark.asyncio
